tests in xmtp_id passing
